[SCH-1233] Fix broken builds after releases (#21)
This PR configures `release-it` to update the tests and examples upon release. This was previously a problem, since both the examples and test contain the version number which would fall out of date. This caused tests to fail after every release (or some jiggering to update these files before tagging+releasing).