feat: add flag to disable run-time validation for gen-js
When generating a JS client, you can now pass in a `--runtimeValidation
<true|false>` flag which toggles whether or not property validation
logic will be generated in the outputted Typewriter client.

This allows you to generate a production version of your Typewriter
client with all validation logic stripped out. This has the benefit
of reducing the bundle size for your end users and increasing the
performance of your analytics (the client just performs a pass-through
to the underlying Segment analytics instance).

BREAKING CHANGE: the run-time flag, `options.propertyValidation`, which
  previously could be used to disable validation at run-time, has been
  removed. It's recommended that you use the build-time flag instead.