docs: Update CONTRIBUTING.md to walkthrough proposing a language (#42)
This PR updates our documentation to give a better walkthrough of how to contribute to Typewriter, so that external (and internal folks not named @gurdasnijor or @colinking) can more easily contribute to Typewriter.

**Updated documentation**: https://github.com/segmentio/typewriter/blob/colin/updated-contributing-docs/.github/CONTRIBUTING.md

Would love to get feedback on this! Is there anything that you're confused by in Typewriter? Do you feel like you could start working on a new language after reading it?