Add e2e test for large integers and floats (#89)
This PR adds a test to catch a regression we saw in our Swift client in v6 of typewriter, where an incorrect memory attribute led to corruption during serialization: https://github.com/segmentio/typewriter/pull/61/files

---

While adding this test, I noticed our iOS clients generate `NSInteger *` which is just a pointer to a primitive `int` under-the-hood. This PR updates the generator to instead us an `NSNumber *` for nullable and optional integers.