Remove unused dependencies (#104)
This PR removes a few unused dependencies and bumps Handlebars to remediate a Snyk vulnerability.

Currently trying to debug why Snyk is failing to parse this valid `package.json` with a `Failed to create project` error. This PR also removes the `"snyk": true` which may be responsible for that. I've removed Snyk from the CI config in favor of the Github integration.